Maui.Graphics: refactor float to double

+++ src/Microsoft.Maui.Graphics/AbstractPattern.cs
@@ -2,12 +2,12 @@ namespace Microsoft.Maui.Graphics
 {
     public abstract class AbstractPattern : IPattern
     {
-        public float Width { get; }
-        public float Height { get; }
-        public float StepX { get; }
-        public float StepY { get; }
+        public double Width { get; }
+        public double Height { get; }
+        public double StepX { get; }
+        public double StepY { get; }
 
-        protected AbstractPattern(float width, float height, float stepX, float stepY)
+        protected AbstractPattern(double width, double height, double stepX, double stepY)
         {
             Width = width;
             Height = height;
@@ -15,11 +15,11 @@ namespace Microsoft.Maui.Graphics
             StepY = stepY;
         }
 
-        protected AbstractPattern(float width, float height) : this(width, height, width, height)
+        protected AbstractPattern(double width, double height) : this(width, height, width, height)
         {
         }
 
-        protected AbstractPattern(float stepSize) : this(stepSize, stepSize)
+        protected AbstractPattern(double stepSize) : this(stepSize, stepSize)
         {
         }
